Move get_form_data into `readable` module
